moved settings and sendmail
