+++ admin/setup.php
@@ -13,7 +13,7 @@ const SETTINGS_PAGE_SLUG = 'tlc-ttsurvey-settings';
 const LOG_PAGE_SLUG = 'tlc-ttsurvey-log';
 
 require_once plugin_path('include/logger.php');
-require_once plugin_path('settings.php');
+require_once plugin_path('include/settings.php');
 
 function handle_admin_init()
 {
