+++ shortcode.php
@@ -10,8 +10,8 @@ if( ! defined('WPINC') ) { die; }
 const LOGIN_FORM_NONCE = 'tlc-ttsurver-login';
 
 require_once plugin_path('include/logger.php');
-require_once plugin_path('settings.php');
-require_once plugin_path('login.php');
+require_once plugin_path('include/settings.php');
+require_once plugin_path('public/login.php');
 
 /**
  * handle login
